requirements.txt
torch>=2.5.1
torchvision>=0.20.1
opencv-python>=4.5.0
numpy>=1.21.0,<2.0.0
Pillow>=8.3.0
trimesh>=3.9.0
pyrr>=0.10.3
scipy>=1.7.0
matplotlib>=3.4.0
tqdm>=4.62.0
transformers>=4.20.0
huggingface-hub>=0.20.0
xatlas>=0.0.9
scikit-image>=0.18.0